requirements.txt
google-generativeai
psycopg2-binary  
Pillow           
sentence-transformers 
python-dotenv
langchain
qdrant-client
tiktoken
pandas
#openeo
#gdal
#rasterio
#scikit-image
#langchain-anthropic
#openai
